'jacocoTestReport' is not part of 'check'
It would trigger all tests (including integration tests). The task
is supposed to run only after all test kinds have been executed.
It should be called explicitly. Usually on CI.
